refactor(auth-mapping): Move authz functions to utils file
